+++ railties/test/isolation/abstract_unit.rb
@@ -144,6 +144,7 @@ module TestHelpers
       end
 
       add_to_config <<-RUBY
+        config.hosts << proc { true }
         config.eager_load = false
         config.session_store :cookie_store, key: "_myapp_session"
         config.active_support.deprecation = :log
@@ -165,6 +166,7 @@ module TestHelpers
       require "action_view/railtie"
 
       @app = Class.new(Rails::Application)
+      @app.config.hosts << proc { true }
       @app.config.eager_load = false
       @app.secrets.secret_key_base = "3b7cd727ee24e8444053437c36cc66c4"
       @app.config.session_store :cookie_store, key: "_myapp_session"